Attempt to get floating panels to work.
